Move DynamicSlice to gpu/primitives
